app/requirements.txt
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.1
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.5
pydantic==2.6.4
pydantic_core==2.16.3
pymarc==5.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.4
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.4
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.29.0
yarl==1.9.4
